Changed from doubles to floats, precision not needed

+++ MathsTools.cpp
@@ -8,17 +8,17 @@ Vector2::Vector2() {
     y = 0;
 }
 
-Vector2::Vector2(double ix, double iy) {
+Vector2::Vector2(float ix, float iy) {
     x = ix;
     y = iy;
 }
 
-Vector2::Vector2(double lst[2]) {
+Vector2::Vector2(float lst[2]) {
     x = lst[0];
     y = lst[1];
 }
 
-Vector2::Vector2(std::vector<double> lst) {
+Vector2::Vector2(std::vector<float> lst) {
     assert(lst.size() == 2);
     x = lst[0];
     y = lst[1];
@@ -38,14 +38,14 @@ Vector2 Vector2::operator-(const Vector2& v1) {
     return result;
 }
 
-Vector2 Vector2::operator*(const double& scale) {
+Vector2 Vector2::operator*(const float& scale) {
     Vector2 result;
     result.x = x * scale;
     result.y = y * scale;
     return result;
 }
 
-Vector2 Vector2::operator/(const double& scale) {
+Vector2 Vector2::operator/(const float& scale) {
     return Vector2(x, y) * (1 / scale);
 }
 
@@ -63,14 +63,14 @@ Vector2 Vector2::operator-=(const Vector2& v1) {
     return result;
 }
 
-Vector2 Vector2::operator*=(const double& scale) {
+Vector2 Vector2::operator*=(const float& scale) {
     Vector2 result = Vector2(x, y) * scale;
     x = result.x;
     y = result.y;
     return result;
 }
 
-Vector2 Vector2::operator/=(const double& scale) {
+Vector2 Vector2::operator/=(const float& scale) {
     Vector2 result = Vector2(x, y) / scale;
     x = result.x;
     y = result.y;
@@ -82,11 +82,11 @@ bool Vector2::operator==(const Vector2& v1)
     return (x == v1.x && y == v1.y);
 }
 
-double Vector2::dotc(const Vector2& v1) {
+float Vector2::dotc(const Vector2& v1) {
     return dot(Vector2(x, y), v1);
 }
 
-double Vector2::abs() {
+float Vector2::abs() {
     return sqrt(x * x + y * y);
 }
 
@@ -111,19 +111,19 @@ Vector3::Vector3() {
     z = 0;
 }
 
-Vector3::Vector3(double ix, double iy, double iz) {
+Vector3::Vector3(float ix, float iy, float iz) {
     x = ix;
     y = iy;
     z = iz;
 }
 
-Vector3::Vector3(double lst[3]) {
+Vector3::Vector3(float lst[3]) {
     x = lst[0];
     y = lst[1];
     z = lst[2];
 }
 
-Vector3::Vector3(std::vector<double> lst) {
+Vector3::Vector3(std::vector<float> lst) {
     assert(lst.size() == 3);
     x = lst[0];
     y = lst[1];
@@ -145,7 +145,7 @@ Vector3 Vector3::operator+(const Vector3& v1) {
     return result;
 }
 
-Vector3 Vector3::operator*(const double& scale) {
+Vector3 Vector3::operator*(const float& scale) {
     Vector3 result;
     result.x = this->x * scale;
     result.y = this->y * scale;
@@ -153,7 +153,7 @@ Vector3 Vector3::operator*(const double& scale) {
     return result;
 }
 
-Vector3 Vector3::operator/(const double& scale) {
+Vector3 Vector3::operator/(const float& scale) {
     return Vector3(x, y, z) * (1 / scale);
 }
 
@@ -173,7 +173,7 @@ Vector3 Vector3::operator-=(const Vector3& v1) {
     return result;
 }
 
-Vector3 Vector3::operator*=(const double& scale) {
+Vector3 Vector3::operator*=(const float& scale) {
     Vector3 result = Vector3(x, y, z) * scale;
     x = result.x;
     y = result.y;
@@ -181,7 +181,7 @@ Vector3 Vector3::operator*=(const double& scale) {
     return result;
 }
 
-Vector3 Vector3::operator/=(const double& scale) {
+Vector3 Vector3::operator/=(const float& scale) {
     Vector3 result = Vector3(x, y, z) / scale;
     x = result.x;
     y = result.y;
@@ -194,7 +194,7 @@ bool Vector3::operator==(const Vector3& v1)
     return (x == v1.x && y == v1.y && z == v1.z);
 }
 
-double Vector3::dotc(const Vector3& v1) {
+float Vector3::dotc(const Vector3& v1) {
     return dot(Vector3(x, y, z), v1);
 }
 
@@ -202,7 +202,7 @@ Vector3 Vector3::crossc(const Vector3& v1) {
     return cross(Vector3(x, y, z), v1);
 }
 
-double Vector3::abs() {
+float Vector3::abs() {
     return sqrt(x * x + y * y + z * z);
 }
 
@@ -228,21 +228,21 @@ Vector4::Vector4() {
     w = 0;
 }
 
-Vector4::Vector4(double ix, double iy, double iz, double iw) {
+Vector4::Vector4(float ix, float iy, float iz, float iw) {
     x = ix;
     y = iy;
     z = iz;
     w = iw;
 }
 
-Vector4::Vector4(double lst[4]) {
+Vector4::Vector4(float lst[4]) {
     x = lst[0];
     y = lst[1];
     z = lst[2];
     w = lst[3];
 }
 
-Vector4::Vector4(std::vector<double> lst) {
+Vector4::Vector4(std::vector<float> lst) {
     assert(lst.size() == 4);
     x = lst[0];
     y = lst[1];
@@ -268,7 +268,7 @@ Vector4 Vector4::operator-(const Vector4& v1) {
     return result;
 }
 
-Vector4 Vector4::operator*(const double& scale) {
+Vector4 Vector4::operator*(const float& scale) {
     Vector4 result;
     result.x = x * scale;
     result.y = y * scale;
@@ -295,7 +295,7 @@ Vector4 Vector4::operator-=(const Vector4& v1) {
     return result;
 }
 
-Vector4 Vector4::operator*=(const double& scale) {
+Vector4 Vector4::operator*=(const float& scale) {
     Vector4 result = Vector4(x, y, z, w) * scale;
     x = result.x;
     y = result.y;
@@ -304,7 +304,7 @@ Vector4 Vector4::operator*=(const double& scale) {
     return result;
 }
 
-Vector4 Vector4::operator/=(const double& scale) {
+Vector4 Vector4::operator/=(const float& scale) {
     Vector4 result = Vector4(x, y, z, w) / scale;
     x = result.x;
     y = result.y;
@@ -318,15 +318,15 @@ bool Vector4::operator==(const Vector4& v1)
     return (x == v1.x && y == v1.y && z == v1.z && w == v1.w);
 }
 
-Vector4 Vector4::operator/(const double& scale) {
+Vector4 Vector4::operator/(const float& scale) {
     return Vector4(x, y, z, w) * (1 / scale);
 }
 
-double Vector4::dotc(const Vector4& v1) {
+float Vector4::dotc(const Vector4& v1) {
     return dot(Vector4(x, y, z, w), v1);
 }
 
-double Vector4::abs() {
+float Vector4::abs() {
     return sqrt(x * x + y * y + z * z + w * w);
 }
 
@@ -345,28 +345,28 @@ Vector4 Vector4::getPerpendicularVector(Vector4 vec1) {
 
 
 // FUNCTION DEFS
-double dot(Vector2 v1, Vector2 v2) {
+float dot(Vector2 v1, Vector2 v2) {
     return ((v1.x * v2.x) + (v1.y * v2.y));
 }
 
-double dot(Vector3 v1, Vector3 v2) {
+float dot(Vector3 v1, Vector3 v2) {
     return ((v1.x * v2.x) + (v1.y * v2.y) + (v1.z*v2.z));
 }
 
-double dot(Vector4 v1, Vector4 v2) {
+float dot(Vector4 v1, Vector4 v2) {
     return ((v1.x * v2.x) + (v1.y * v2.y) + (v1.z*v2.z) + (v1.w + v2.w));
 }
 
 
-double getAngle(Vector2 v1, Vector2 v2) {
+float getAngle(Vector2 v1, Vector2 v2) {
     return acos(dot(v1, v2) / (v1.abs(), v2.abs()));
 }
 
-double getAngle(Vector3 v1, Vector3 v2) {
+float getAngle(Vector3 v1, Vector3 v2) {
     return acos(dot(v1, v2) / (v1.abs(), v2.abs()));
 }
 
-double getAngle(Vector4 v1, Vector4 v2) {
+float getAngle(Vector4 v1, Vector4 v2) {
     return acos(dot(v1, v2) / (v1.abs(), v2.abs()));
 }
 
@@ -381,16 +381,16 @@ Vector3 cross(Vector3 v1, Vector3 v2) {
 
 
 Vector2 normaliseV2(Vector2 vec) {
-    double absolute = vec.abs();
+    float absolute = vec.abs();
     return Vector2(vec.x / absolute, vec.y / absolute);
 }
 
 Vector3 normaliseV3(Vector3 vec) {
-    double absolute = vec.abs();
+    float absolute = vec.abs();
     return Vector3(vec.x / absolute, vec.y / absolute, vec.z / absolute);
 }
 
 Vector4 normaliseV4(Vector4 vec) {
-    double absolute = vec.abs();
+    float absolute = vec.abs();
     return Vector4(vec.x / absolute, vec.y / absolute, vec.z / absolute, vec.w / absolute);
 }
